# receive: use s2 when decompressing snappy data, reduce CPU usage for incoming remote write requests by ~5-7%