@GiedriusS 2021-06-18
The ASM code seems to roughly follow the Go code. So, it is easier to compare https://github.com/klauspost/compress/blob/master/s2/decode_other.go with https://github.com/golang/snappy/blob/master/decode_other.go. `diff -u ...` makes it clear where time is saved. (: I am not an expert on compression but it seems like time is saved by essentially decoding more at once in each loop.